Fix a bug

+++ vars/deleteWorkspace.groovy
@@ -3,15 +3,24 @@ import java.io.File
 
 /**
  * Todo -
- *         Users may want to keep their workspace, but deleting their
- *         runtime env is a must, so we have to provide two options:
- *                  Remove Ws -> Delete entire path
- *                  Keep Ws -> Delete .JD_CODE_BUILD only
+ *        Require more test on whether folder exists / available
  *
  */
-def call(def path){
+def call(def path,def depth="ALL"){
+
     assert path.length()>0 :"Empty path given"
-    def ws = new File(path)
+    def entrance = new File(path)
+    if (!entrance.exists()) {
+        return
+    }
+
     println("Cleaning workspace..")
-    ws.deleteDir()
+
+    if(depth == "ALL") {
+        def ws = new File(path + "/workspace")
+        ws.deleteDir()
+    }
+
+    def meta = new File(path + "/.JD_CODE_BUILD")
+    meta.deleteDir()
 }
